Move the visualiser code into fe-analyselayout.
